% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_coastline.R
\name{annotate_coastline}
\alias{annotate_coastline}
\title{Annotate coastline}
\usage{
annotate_coastline(city = c("Melbourne", "Sydney", "Brisbane"),
  coastline.color = theGrey)
}
\arguments{
\item{city}{Which city is to be annotated with its coastline?}

\item{coastline.color}{Colour of the path.}
}
\description{
Used in a \code{ggplot2} call to add Sydney or Melbourne's
}
\examples{
\dontrun{
ggplot(...) + 
  geom_polygon() +
  annotate_coastline("Sydney")
}
}
