% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_chart_data.R
\name{bundle_chart_data}
\alias{bundle_chart_data}
\title{Bundle chart data}
\usage{
bundle_chart_data(chart_data_table,
  chart_data.xlsx = "atlas/bundled-chart-data.xlsx")
}
\arguments{
\item{chart_data_table}{A data.table including a Figure C.M column (\code{fig_no}) and the file name for the 
figure image (\code{extract}).}

\item{chart_data.xlsx}{Excel file to write to.}
}
\description{
Bundle chart data
}
