% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-coastlines.R
\docType{data}
\name{coastlines}
\alias{coastlines}
\alias{Melbourne_coastline}
\alias{Sydney_coastline}
\title{Coastline coordinates}
\format{\describe{
\item{\code{long}}{Longitude of coastline.}
\item{\code{lat}}{Latitude of coastline.}
\item{\code{order}}{Order of polygon.}
}}
\usage{
Melbourne_coastline

Sydney_coastline
}
\description{
Used to annotate maps to provide the coastline around cities.
}
\keyword{datasets}
