% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattanCharts-package.R
\docType{package}
\name{grattanCharts-package}
\alias{grattanCharts}
\alias{grattanCharts-package}
\title{Colours, charts, and other common quant tasks}
\description{
Grattan package
}
\details{
Colours, charts, and other common quant tasks
}
\author{
\email{hugh.parsonage+grattanpackage@grattan.edu.au}
}
\keyword{package}

