% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_knitting.R
\name{if_knitting}
\alias{if_knitting}
\title{Run code conditional on knitting}
\usage{
if_knitting(yes_expr, no_expr)
}
\arguments{
\item{yes_expr, no_expr}{Expressions to run if knitr is running or if it is not running.}
}
\description{
Run code conditional on knitting
}
