% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_reporters.R
\name{install_reporters}
\alias{install_reporters}
\title{Install the dependencies for powerpoints}
\usage{
install_reporters()
}
\description{
Installs the now-deprecated dependencies necessary to save reasonable-fidelity
graphics into PowerPoint.
}
