% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-dimensions.R
\docType{data}
\name{knitr.dimensions}
\alias{knitr.dimensions}
\alias{fig_height}
\alias{fig_width}
\alias{out_height}
\alias{out_width}
\alias{column_width}
\title{Dimensions for knitr}
\format{An object of class \code{numeric} of length 1.}
\usage{
fig_height

fig_width

out_height

out_width

column_width
}
\description{
Dimensions for knitr. Made user-visible for chunk-specific options.
}
\keyword{datasets}
