% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pptx.R
\name{save_pptx}
\alias{save_pptx}
\title{Save plot as PowerPoint}
\usage{
save_pptx(p, filename, template = c("presentation", "report"),
  template.file = NULL)
}
\arguments{
\item{p}{A \code{ggplot2} object.}

\item{filename}{Filename for the PowerPoint file.}

\item{template}{Either \code{presentation} or \code{report}, the template on Dropbox to be used. 
Only useful if connected to Dropbox. Can be overriden by \code{template.file}.}

\item{template.file}{If set, the file (must have extension \code{.pptx}) to be used in the template.}
}
\description{
Save plot as PowerPoint
}
