% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_rnw.R
\name{start_rnw}
\alias{start_rnw}
\title{Create a knitr document to generate charts}
\usage{
start_rnw(path, rstudio = TRUE)
}
\arguments{
\item{path}{A file path, either to a filename to create, or a directory into which to create a file.
Will be assumed to be a directory, unless it ends in \code{.Rnw} in which case it will be taken
to be the filename.}
}
\description{
Not to be used for reports: clone the grattex repository instead. Use-case is
creating an atlas of charts.
}
\details{
If using RStudio, will pop to the document when completed.
}
