% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pptx.R
\name{save_pptx}
\alias{save_pptx}
\title{Save plot as PowerPoint}
\usage{
save_pptx(p, filename, template = c("presentation", "report"),
  template.file = NULL)
}
\arguments{
\item{p}{A \code{ggplot2} object.}

\item{filename}{Filename for the PowerPoint file.}

\item{template}{Either \code{presentation} or \code{report}, the template on Dropbox to be used. 
Only useful if connected to Dropbox. Can be overriden by \code{template.file}.}

\item{template.file}{If set, the file (must have extension \code{.pptx}) to be used in the template.}
}
\description{
Save plot as PowerPoint
}
\details{
Generally, you will need the \code{ReporteRs} package which has been deprecated on CRAN, 
so you will need to install it remotely. \code{\link{install_reporters}}.
}
\examples{
\dontrun{
source("https://gist.githubusercontent.com/HughParsonage/60581c4595f3f10ba201faeacc46cca5/raw/31316307432ee6fef3164e212d21f173ec49d25f/employment-QLD")
}

}
