% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_knitr_for_grattex.R
\name{setup_knitr_for_grattex}
\alias{setup_knitr_for_grattex}
\title{Setup knitr for Grattan reports}
\usage{
setup_knitr_for_grattex(chunktimings.txt = "CHUNKTIMINGS.txt",
  bundle_chart_data, theGlobalEnv = .GlobalEnv, error = FALSE)
}
\arguments{
\item{chunktimings.txt}{The file to write chunk timings}

\item{bundle_chart_data}{Whether or not to bundle chart data.}

\item{theGlobalEnv}{Must be .GlobalEnv.}

\item{error}{Passed to \code{knitr::opts_chunk$set(error)}.}
}
\value{
Called for its side-effect: when knitr is in progress, sets up
knitr for consistent formatting.
}
\description{
Setup knitr for Grattan reports
}
